requirements.txt
python-telegram-bot[job-queue]>=20.6
python-dotenv